rename codeGenerator.py to code_generator.py
